Implement grapple swinging
This grapple swinging is similar to Spiderman's web swinging where
the force applied to the player is based on the player's velocity
and where the player is during the swing.

A BPI_Interaction interface is used for ray casting for an
interactable actor and the player will have to look at the anchor
to cast the grapple.