Merge the grapple hook into the VR pawn
The grapple hook implementations are merged into the default VR
pawn under VRTemplate/Blueprints/VRPawn. The parent class is changed
to be Character instead of Pawn so that all movement that affects
the Character Movement component in the previous blueprint would
still work as intended.

The Use Controller Rotation Pitch is enabled in the VR Pawn
blueprint so that the mouse input can rotate the camera's pitch
for testing.

Let's:
- Disable the pitch rotation during VR testing.
- Convert keyboard and mouse inputs into VR inputs.
- Ray cast for interactable actors from the VR controller instead of
the VR camera.