Upgrade the swinging motion
The swinging will now apply a total of 3 forces, the
radial force, tangential force, and damping tangential force. The
player velocity is also dampened during collisions with other
actors.

The player will also swing around the impact point instead of the
attached actor's center point which was not intended. The map is
also cleared out to test the swinging.

An initial push force is applied if the player is on ground to
initiate the pendulum motion.

This should all make a smoother and more realistic swinging motion
and experience for the player.