Add hooking ability to grapple component
The player will be able to choose between swinging across or
hooking towards the interaction point when interact button is
pressed.

This allows more flexible movement gameplay that follows our
initial intention. This also means that we do not have to create
another interface or a separate input apart from swinging, for
hooking, instead, everything still relies on the interact button,
and only swingable points are hook-able.

A boolean field "isHooking" is added to VRPawn to prevent compile
errors on the grapple component. This field is used for toggling
between the grappling modes.

Let's:
- create a toggle on the middle finger switch to toggle between
swinging and hooking modes.
- optionally add an interface to determine grapple-able actors.